Support exporting image of more formats

+++ src/utilities.ts
@@ -6,7 +6,7 @@ export const readFileAsync = util.promisify(fs.readFile);
 export const writeFileAsync = util.promisify(fs.writeFile);
 
 function joinBuffers(buffers: Uint8Array[]): string {
-    return Buffer.concat(buffers).toString()
+    return Buffer.concat(buffers).toString();
 }
 
 export function runChildProcess(
@@ -14,7 +14,7 @@ export function runChildProcess(
     args: string[],
     cwd: string,
     input: string,
-    cancel: Promise<void>
+    cancel?: Promise<void>
 ): Promise<[number, string, string]> {
     return new Promise((resolve, reject) => {
         const process = child_process.spawn(program, args, { cwd });
@@ -26,7 +26,9 @@ export function runChildProcess(
         process.stdout.on("data", (chunk) => stdoutBuffer.push(chunk));
         process.stderr.on("data", (chunk) => stderrBuffer.push(chunk));
 
-        cancel.then(() => process.kill());
+        if (cancel) {
+            cancel.then(() => process.kill());
+        }
 
         try {
             process.stdin.end(input);
